chore(profile): Move widely used components out of Profile view
