+++ src/views/Home/components/UserBanner/UserDetail.tsx
@@ -3,7 +3,7 @@ import { NoProfileAvatarIcon, Flex, Heading, Skeleton, Text, Box } from '@pancak
 import { useWeb3React } from '@web3-react/core'
 import styled from 'styled-components'
 import { useProfile } from 'state/profile/hooks'
-import ProfileAvatar from 'views/Profile/components/ProfileAvatar'
+import ProfileAvatarWithTeam from 'components/ProfileAvatarWithTeam'
 import { useTranslation } from 'contexts/Localization'
 import truncateHash from 'utils/truncateHash'
 
@@ -69,7 +69,7 @@ const UserDetail = () => {
     <>
       <Desktop>
         <Box mr="24px">
-          <Sticker>{profile ? <ProfileAvatar profile={profile} /> : <StyledNoProfileAvatarIcon />}</Sticker>
+          <Sticker>{profile ? <ProfileAvatarWithTeam profile={profile} /> : <StyledNoProfileAvatarIcon />}</Sticker>
         </Box>
         <Flex flexDirection="column">
           {getDesktopHeading()}
