+++ src/views/Ifos/components/IfoFoldableCard/IfoPoolCard/ContributeModal.tsx
@@ -8,7 +8,7 @@ import { PoolIds, Ifo } from 'config/constants/types'
 import { WalletIfoData, PublicIfoData } from 'views/Ifos/types'
 import { useTranslation } from 'contexts/Localization'
 import { getBalanceAmount } from 'utils/formatBalance'
-import ApproveConfirmButtons from 'views/Profile/components/ApproveConfirmButtons'
+import ApproveConfirmButtons from 'components/ApproveConfirmButtons'
 import useApproveConfirmTransaction from 'hooks/useApproveConfirmTransaction'
 import { useCallWithGasPrice } from 'hooks/useCallWithGasPrice'
 import { DEFAULT_TOKEN_DECIMAL } from 'config'
