+++ src/views/Profile/ProfileCreation/Mint.tsx
@@ -9,9 +9,9 @@ import { Nft } from 'config/constants/nfts/types'
 import { FetchStatus, useGetCakeBalance } from 'hooks/useTokenBalance'
 import nftList from 'config/constants/nfts'
 import { useCallWithGasPrice } from 'hooks/useCallWithGasPrice'
+import ApproveConfirmButtons from 'components/ApproveConfirmButtons'
 import SelectionCard from '../components/SelectionCard'
 import NextStepButton from '../components/NextStepButton'
-import ApproveConfirmButtons from '../components/ApproveConfirmButtons'
 import useProfileCreation from './contexts/hook'
 import { MINT_COST, STARTER_BUNNY_IDENTIFIERS } from './config'
