+++ src/views/Profile/components/ConfirmProfileCreationModal.tsx
@@ -10,8 +10,8 @@ import { fetchProfile } from 'state/profile'
 import useToast from 'hooks/useToast'
 import { useCallWithGasPrice } from 'hooks/useCallWithGasPrice'
 import { ToastDescriptionWithTx } from 'components/Toast'
+import ApproveConfirmButtons from 'components/ApproveConfirmButtons'
 import { REGISTER_COST } from '../ProfileCreation/config'
-import ApproveConfirmButtons from './ApproveConfirmButtons'
 import { State } from '../ProfileCreation/contexts/types'
 
 interface Props {
