+++ src/views/TradingCompetition/components/RegisterModal/index.tsx
@@ -2,11 +2,11 @@ import React from 'react'
 import styled from 'styled-components'
 import { Modal, Button, NoProfileAvatarIcon, Flex } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
+import ProfileAvatarWithTeam from 'components/ProfileAvatarWithTeam'
 import { CompetitionProps } from '../../types'
 import MakeProfile from './MakeProfile'
 import ReactivateProfile from './ReactivateProfile'
 import RegisterWithProfile from './RegisterWithProfile'
-import ProfileAvatar from '../../../Profile/components/ProfileAvatar'
 
 const AvatarWrapper = styled.div`
   height: 64px;
@@ -40,7 +40,9 @@ const RegisterModal: React.FC<CompetitionProps> = ({ onDismiss, profile, onRegis
   return (
     <Modal title={t('Register')} onDismiss={onDismiss}>
       <Flex flexDirection="column" alignItems="center" maxWidth="400px">
-        <AvatarWrapper>{profile ? <ProfileAvatar profile={profile} /> : <StyledNoProfileAvatarIcon />}</AvatarWrapper>
+        <AvatarWrapper>
+          {profile ? <ProfileAvatarWithTeam profile={profile} /> : <StyledNoProfileAvatarIcon />}
+        </AvatarWrapper>
         {modalInner()}
       </Flex>
       <Button variant="text" onClick={onDismiss}>
